// frontend/src/lib/types/stepsTypes.ts
export type StepType =
  | "CreateFile"
  | "CreateFolder"
  | "EditFile"
  | "DeleteFile"
  | "RunScript";

export interface Step {
  id: number;
  title: string;
  description: string;
  type: StepType;
  status: "pending" | "in-progress" | "completed";
  code?: string;
  path?: string;
}

export interface Project {
  prompt: string;
  steps: Step[];
}

export interface FileItem {
  path: string;
  content: string;
  language: string;
}

export interface FileViewerProps {
  file: FileItem | null;
  onClose: () => void;
}